receiving data
receives selected relatives from Relative.js and also the id of the calling resident (to be used in back-arrow), however since i changed the links the page will only display when typing in the url manually or when refreshing the page. from the rendered page it successfully goes back to Relative.js with the chosen resident_id. in other words the only issue seems to be rendering the page immediately as it is linked to in Relative.js. as you cant get to the page from Relative.js anymore, the selected relatives will never be shown (line 17 will be false), but i tested so that data is passed on properly before i started tinkering with the links. as soon as the linking is fixed, the selected relatives should be displayed as intended.