redirects to feedback page after updating resident
note: actually updating a resident still lacks support. tested using placeholder function without any actual requests, then commented the form away again to avoid misleading the user (as the functionality doesnt actually exist). however, as soon as the appropriate requests are added and the form is un-hidden the code should redirect and give you feedback as is appropriate (in other words the rest of the code is done, its just the updating itself that is missing)